+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## master (unreleased)
 	
 ### New Features 
-            
+
+    - Implement :block-forms to allow the user to include all forms with the exception of some
+    
 ### Changes
 
     - Improved tracing perf